==> 02 - Scraping Instagram - stats from all the pictures.ipynb <==
from bs4 import BeautifulSoup
from bs4 import SoupStrainer
import pandas as pd
import numpy as np
import urllib
import urllib.parse
import re
import os
import subprocess
import requests
import random
import time
from requests import get

account =  ['resortalyeska',
            'skiheavenly',
            'kirkwoodmtn',
            'mammothmountain',
            'northstar_california',
            'squawalpine',
            'sugarbowlresort',
            'arapahoe_basin',
            'aspensnowmass',
            'aspenco',
            'beavercreek',
            'breckenridgemtn',
            'coppermtn',
            'skicrestedbutte',
            'keystone_resort',
            'lovelandskiarea',
            'skipowderhorn',
            'silvertonmtn',
            'steamboatresort',
            'tellurideski',
            'vailmtn',
            'winterparkresort',
            'schweitzer_mountain',
            'sunvalley',
            'tamarackresort',
            'bigskyresort',
            'bridgerbowl',
            'mtroseskitahoe',
            'skitaos',
            'mthoodmeadows',
            'altaskiarea',
            'brightonresort',
            'deervalleyresort',
            'pcski',
            'powdermountain',
            'snowbasinresort',
            'snowbird',
            'solitudemtn',
            'jaypeakresort',
            'killingtonmtn',
            'madriverglen',
            'stowemt',
            'crystalmountain',
            'mtbakerskiarea',
            'stevenspass',           
            'grandtargheeresort',
            'jacksonhole',
            'sundanceresort',
            'bogusbasin',
            'summitatsnoqualmie',
            'skipurg',
            'brianheadresort',
            'sierra_at_tahoe',
            'skieaglecrest',
            'silvermountain',
            'skisantafe',
            'skithebeav',
            'brundagemtn',
            'skieaglepoint',
            'whitepass',
            'losttrailskiarea',
            'smugglersnotchvt',
            'redlodge',
            'monarchmountain',
            'sugarbush_vt',
            'skiwhitefish',
            'wolfcreekski',
            'mtbachelor',
            'whistlerblackcomb',
            'revelstoke',
            'ferniealpineresort',
            'whitewaterskiresort',
            'redresort',
            'sunshinevillage',
            'kickinghorsemtn',
            'powderkingmountainresort',
            'skilouise',
            'panoramaresort',
            'castlemountainresort',
            'skibigwhite',
            'cypressmtn',
            'silverstarmtnresort',
            'sunpeaksresort',
            'mountwashington',
            'timberlinelodge',
            'skihomewood',
            'eldoramtnresort',
            'bearvalleymountainupdates',
            'diamondpeak',
            '49degreesnorth',
            'lookoutpass',
            'discoveryskiarea',
            'pebbleskiarea',
            'sunlightmountainresort',
            'skibowl',
            'dodgeridge',
            'anthonylakes',
            'marmotbasin',
            'apexmtnresort']

for i in range(0, len(account)):
    
    # open the VPN-connection and wait for 60 seconds for the program to actually connect
    vpn = subprocess.Popen('c:\\Program Files\\Trust.Zone VPN Client\\trustzone_x64.exe')
    time.sleep(60)

    # collect and print the public IP address to make sure the VPN worked
    public_ip = get('https://api.ipify.org').text
    print('The public IP address is:',public_ip)
    
    # set the URL to the Instagram account
    url = 'https://www.instagram.com/' + str(account[i])

    # make a folder to save the data
    path = urllib.parse.urlparse(url).path
    folder_name = path.replace('/', '')
    folder_path = 'images/{}/'.format(folder_name)

    account_data = pd.read_csv(os.path.join(folder_path, str(account[i]) + "_pictures.csv"), sep = ',')
    number_input = account_data.shape[0]
    
    image_url_list = account_data.values.tolist()
    
    only_meta_tags = SoupStrainer('meta')

    image_info_list = []
    image_info = ()
    image_counter = 0

    for j in range(len(image_url_list)):
        video_link = str()
        image_link = str()
        likes_image = str()
        comments_image = str()
        hashtags_image = 0

        page = requests.get(image_url_list[j][2])
        soup_image = BeautifulSoup(page.text, 'html.parser', parse_only = only_meta_tags)

        for link in soup_image.find_all('meta'):
            metaprop = link.get('property', '')

            if metaprop == "og:video":
                video_link = link.get('content')

            if metaprop == "og:image":
                image_link = link.get('content')

            if metaprop == "og:description":
                description_info = link.get('content')

                # parse the number of likes (get all the text before Likes or vind-ik-leuks)
                likes_image_str = re.search('(\d.+)?(?=Likes|vind-ik-leuks)', str(description_info)).group(0).strip()

                # get rid of the k or K (thousands) and commas and convert to integers
                likes_image_temp = re.sub('[kK]+', '', likes_image_str)
                if likes_image_temp == likes_image_str:
                    likes_image = int(float(re.sub('[,|.]+', '', likes_image_temp)))
                else:
                    likes_image = int(float(re.sub('[,|.]+', '.', likes_image_temp)) * 1000)
                
                # parse the number of comments (get all the text between the comma and Comments or reacties)
                comments_image_str = re.search('(?:s,)(.*?)(?=Comments|reacties)', str(description_info)).group(0).strip()[3:]
                
                # get rid of the k or K (thousands) and commas and convert to integers
                comments_image_temp = re.sub('[kK]+', '', comments_image_str)
                if comments_image_temp == comments_image_str:
                    comments_image = int(float(re.sub('[,|.]+', '', comments_image_temp)))
                else:
                    comments_image = int(float(re.sub('[,|.]+', '.', comments_image_temp)) * 1000)

            if metaprop == "og:title":
                title_info = link.get('content')
                date_post = re.search('(Jan|Feb|Mar|Apr|May|Jun|Jul|Aug|Sep|Oct|Nov|Dec).(\d.+)(?=at)', str(title_info)).group(0).strip()
                time_post = re.search('(\d\sat.*$)', str(title_info)).group(0).strip()[5:]
            if metaprop == "instapp:hashtags":
                hashtags_image += 1
                
        image_counter += 1
        account_id = i + 1
        image_id = (i + 1) * 100000 + image_counter
        followers_account = image_url_list[j][3]
        date_scrape = time.strftime("%b %#d, %Y")
        image_info = (account_id, image_id, followers_account, title_info, date_post, time_post, date_scrape, description_info, likes_image, comments_image, hashtags_image, video_link, image_link)
        image_info_list.append(image_info)
    
        if image_counter%100 == 0:
            print('Stats of', image_counter, 'images processed')
    
        # wait after every picture between 0 and 2 seconds to minimize the chance of being banned
        # wait after every 100 pictures for period between 10 and 20 seconds to minimize the chance of being banned
        if image_counter%100 == 0:
            period_100 = np.random.uniform(10, 20)
            time.sleep(period_100)
        else:
            period_1 = np.random.uniform(0, 2)
            time.sleep(period_1)
            
    likes_comments = pd.DataFrame([[x[0], x[1], x[2], x[3], x[4], x[5], x[6], x[7], x[8], x[9], x[10], x[11], x[12]] for x in image_info_list])
    likes_comments.columns = ['account_id', 'image_id', 'followers_account', 'title_info', 'date_post', 'time_post', 'date_scrape', 'description', 'likes_image', 'comments_image', 'hashtags_image', 'video_link', 'image_link']
    likes_comments.to_csv(os.path.join(folder_path, str(account[i]) + "_stats.csv"), sep = ',', index = False)
    
    number_output = likes_comments.shape[0]
    
    print('Number of records of input and output are the same:', number_input == number_output)
    
    vpn.terminate()
    time.sleep(60)
