==> 11 - Random sample image classification.ipynb <==
import pandas as pd
import numpy as np
import random
import os
from shutil import copy

account_data_01 = pd.read_csv('results/dataset_analysis.csv', low_memory = False)
clusters = pd.read_csv('results/clusters.csv', low_memory = False)
account_data_02 = account_data_01.merge(clusters, on = 'image_id', how = 'inner')

image_ids_train = pd.read_csv('results/image_ids_train.csv', low_memory = False)
image_ids_val = pd.read_csv('results/image_ids_val.csv', low_memory = False)
image_ids_test = pd.read_csv('results/image_ids_test.csv', low_memory = False)

training_df = account_data_02.merge(image_ids_train, on = 'image_id', how = 'inner')
training_df = training_df[['likes_groups', 'cluster', 'accountname', 'image_id']]

validation_df = account_data_02.merge(image_ids_val, on = 'image_id', how = 'inner')
validation_df = validation_df[['likes_groups', 'cluster', 'accountname', 'image_id']]

test_df = account_data_02.merge(image_ids_test, on = 'image_id', how = 'inner')
test_df = test_df[['likes_groups', 'cluster', 'accountname', 'image_id']]

print('Number of samples in training_df:', training_df.shape[0])
print('Number of samples in validation_df:', validation_df.shape[0])
print('Number of samples in test_df:', test_df.shape[0])

random.seed(42)
sample_ids = random.sample(account_data_02[['likes_groups', 'cluster', 'accountname', 'image_id']].values.tolist(), 5000)

picture_size = 256

for j in range(0, len(sample_ids)):

    file = 'images/' + str(sample_ids[j][2]) + '/crop/' + str(sample_ids[j][3]) + '_' + str(picture_size) + '.jpg'
    destination = 'image_classifier/'
    
    if not os.path.exists(destination):
        os.makedirs(destination)
    
    # copy the picture if it exists
    if os.path.isfile(file):
        copy(file, destination)
